CAS-1116 Fix bug in writing integer value caused by typo.

+++ cas-server-integration-memcached/src/main/java/org/jasig/cas/ticket/registry/support/kryo/serial/MultiTimeUseOrTimeoutExpirationPolicySerializer.java
@@ -38,7 +38,7 @@ public class MultiTimeUseOrTimeoutExpirationPolicySerializer extends SimpleSeria
     }
 
     public void write(final ByteBuffer buffer, MultiTimeUseOrTimeoutExpirationPolicy policy) {
-        buffer.putLong((Integer) fieldHelper.getFieldValue(policy, "numberOfUses"));
+        buffer.putInt((Integer) fieldHelper.getFieldValue(policy, "numberOfUses"));
         buffer.putLong((Long) fieldHelper.getFieldValue(policy, "timeToKillInMilliSeconds"));
     }
